requirements.txt
langchain
huggingface_hub
chainlit
python-dotenv

